python_utils/requirements.txt
accelerate==0.33.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.30.2
azure-identity==1.17.1
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.3
build==1.2.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==43.0.0
dataclasses-json==0.6.7
decorator==5.1.1
demjson3==3.0.6
Deprecated==1.2.14
deprecation==2.1.0
diffusers==0.30.1
dirtyjson==1.0.8
distro==1.9.0
docstring-parser==0.15
docx2txt==0.8
einops==0.8.0
exceptiongroup==1.1.1
fastapi==0.112.2
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
google-auth==2.34.0
googleapis-common-protos==1.56.4
greenlet==3.0.3
grpcio==1.66.0
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.5
httptools==0.6.1
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
importlib_metadata==8.0.0
importlib_resources==6.4.4
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
kubernetes==30.1.0
lancedb==0.3.6
llama-cloud==0.0.15
llama-index==0.10.68
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.68.post1
llama-index-embeddings-azure-openai==0.1.11
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48.post3
llama-index-llms-azure-openai==0.1.10
llama-index-llms-openai==0.1.31
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.0
openai==1.42.0
opencv-python==4.10.0.84
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.7
overrides==7.7.0
packaging==23.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.0.0
portalocker==2.10.1
posthog==3.5.2
prettytable==3.11.0
prompt-toolkit==3.0.36
protobuf==3.20.0
psutil==6.0.0
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.8.21
pymemgpt==0.3.24
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pytest==7.3.1
python-box==7.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.4
pywin32==306
PyYAML==6.0.2
questionary==2.0.1
ratelimiter==1.2.0.post0
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.1
rsa==4.9
safetensors==0.4.4
semver==3.0.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
sqlalchemy-json==0.7.0
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.16
starlette==0.38.2
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
tiktoken==0.5.2
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
torchvision==0.19.0
tqdm==4.66.5
transformers==4.25.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
voluptuous==0.15.2
watchfiles==0.23.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.0
Werkzeug==3.0.4
wrapt==1.16.0
xformers==0.0.27.post2
yarl==1.9.4
zipp==3.20.0
